Make remaining custom operator new size params size_t for x64 (x64 fix)
Change the size argument of the five remaining class/placement
operator new overloads from uint to size_t so template deduction
and the standard operator-new signature hold on x64 (LLP64):
CActivity, CInstance, CPartitionSpace, CLRU, CScheduler. Scheduler's
u_alignment stays uint. Value/behaviour unchanged on Win32 (size_t
== uint there). Win32 default (VER_ASM=TRUE) build green; x64
(VER_ASM=FALSE) diagnostic build errors 1037 -> 736, C2821 218 -> 28.

+++ jp2_pc/Source/Game/AI/Activity.cpp
@@ -143,7 +143,7 @@ extern CProfileStat psPathfinder;
 
 
 	//*****************************************************************************************
-	void* CActivity::operator new(uint  u_size_type)
+	void* CActivity::operator new(size_t  u_size_type)
 	{
 #ifdef LOG_MEM
